+++ app/gui2/src/components/GraphEditor/collapsing.ts
@@ -1,9 +1,10 @@
 import { GraphDb } from '@/stores/graph/graphDatabase'
+import { assert } from '@/util/assert'
 import { Ast } from '@/util/ast'
 import { moduleMethodNames } from '@/util/ast/abstract'
+import { nodeFromAst } from '@/util/ast/node'
 import { unwrap } from '@/util/data/result'
 import { tryIdentifier, type Identifier } from '@/util/qualifiedName'
-import assert from 'assert'
 import * as set from 'lib0/set'
 import { IdMap, type ExprId } from 'shared/yjsModel'
 
@@ -39,7 +40,7 @@ interface Output {
 interface RefactoredInfo {
   /** The id of the refactored node. */
   id: ExprId
-  /** The pattern of the refactored node. Included for convinience, collapsing does not affect it. */
+  /** The pattern of the refactored node. Included for convenience, collapsing does not affect it. */
   pattern: string
   /** The list of necessary arguments for a call of the collapsed function. */
   arguments: Identifier[]
@@ -88,12 +89,13 @@ export function prepareCollapsedInfo(selected: Set<ExprId>, graphDb: GraphDb): C
   // If there is no output found so far, it means that none of our nodes is used outside
   // the extracted function. In such we will return value from arbitrarily chosen leaf.
   if (output == null) {
-    const arbitaryLeaf = set.first(leaves)
-    if (arbitaryLeaf == null) throw new Error('Cannot select the output node, no leaf nodes found.')
-    const outputNode = graphDb.nodeIdToNode.get(arbitaryLeaf)
-    if (outputNode == null) throw new Error(`The node with id ${arbitaryLeaf} not found.`)
+    const arbitraryLeaf = set.first(leaves)
+    if (arbitraryLeaf == null)
+      throw new Error('Cannot select the output node, no leaf nodes found.')
+    const outputNode = graphDb.nodeIdToNode.get(arbitraryLeaf)
+    if (outputNode == null) throw new Error(`The node with id ${arbitraryLeaf} not found.`)
     const identifier = unwrap(tryIdentifier(outputNode.pattern?.code() || ''))
-    output = { node: arbitaryLeaf, identifier }
+    output = { node: arbitraryLeaf, identifier }
   }
 
   const pattern = graphDb.nodeIdToNode.get(output.node)?.pattern?.code() ?? ''
@@ -131,56 +133,74 @@ function findSafeMethodName(module: Ast.Module, baseName: string): string {
 const MODULE_NAME = 'Main'
 const COLLAPSED_FUNCTION_NAME = 'collapsed'
 
+interface CollapsingResult {
+  /** The ID of the node refactored to the collapsed function call. */
+  refactoredNodeId: ExprId
+  /** IDs of nodes inside the collapsed function, except the output node.
+   * The order of these IDs is reversed comparing to the order of nodes in the source code.
+   */
+  collapsedNodeIds: ExprId[]
+  /** ID of the output node inside the collapsed function. */
+  outputNodeId?: ExprId | undefined
+}
+
 /** Perform the actual AST refactoring for collapsing nodes. */
 export function performCollapse(
   info: CollapsedInfo,
-  module: Ast.Module,
+  edit: Ast.MutableModule,
   topLevel: Ast.BodyBlock,
   db: GraphDb,
   currentMethodName: string,
-): Ast.MutableModule {
-  const functionAst = Ast.findModuleMethod(module, currentMethodName)
+): CollapsingResult {
+  const functionAst = Ast.findModuleMethod(edit, currentMethodName)
   if (!(functionAst instanceof Ast.Function) || !(functionAst.body instanceof Ast.BodyBlock)) {
     throw new Error(`Expected a collapsable function, found ${functionAst}.`)
   }
   const functionBlock = functionAst.body
-  const posToInsert = findInsertionPos(module, topLevel, currentMethodName)
-  const collapsedName = findSafeMethodName(module, COLLAPSED_FUNCTION_NAME)
+  const posToInsert = findInsertionPos(edit, topLevel, currentMethodName)
+  const collapsedName = findSafeMethodName(edit, COLLAPSED_FUNCTION_NAME)
   const astIdsToExtract = new Set(
     [...info.extracted.ids].map((nodeId) => db.nodeIdToNode.get(nodeId)?.outerExprId),
   )
   const astIdToReplace = db.nodeIdToNode.get(info.refactored.id)?.outerExprId
   const collapsed = []
   const refactored = []
-  const edit = module.edit()
   const lines = functionBlock.lines()
+  const { ast: refactoredAst, nodeId: refactoredNodeId } = collapsedCallAst(
+    info,
+    collapsedName,
+    edit,
+  )
   for (const line of lines) {
     const astId = line.expression?.node.exprId
-    const ast = astId != null ? module.get(astId) : null
+    const ast = astId != null ? edit.get(astId) : null
     if (ast == null) continue
     if (astIdsToExtract.has(astId)) {
       collapsed.push(ast)
       if (astId === astIdToReplace) {
-        const newAst = collapsedCallAst(info, collapsedName, edit)
-        refactored.push({ expression: { node: newAst } })
+        refactored.push({ expression: { node: refactoredAst } })
       }
     } else {
       refactored.push({ expression: { node: ast } })
     }
   }
+  const collapsedNodeIds = collapsed.map((ast) => nodeFromAst(ast).rootSpan.exprId).reverse()
+  let outputNodeId: ExprId | undefined
   const outputIdentifier = info.extracted.output?.identifier
   if (outputIdentifier != null) {
-    collapsed.push(Ast.Ident.new(edit, outputIdentifier))
+    const ident = Ast.Ident.new(edit, outputIdentifier)
+    collapsed.push(ident)
+    outputNodeId = ident.exprId
   }
-  // Update the definiton of refactored function.
+  // Update the definiton of the refactored function.
   const refactoredBlock = Ast.BodyBlock.new(refactored, edit)
   edit.replaceRef(functionBlock.exprId, refactoredBlock)
-  // new Ast.BodyBlock(edit, functionBlock.exprId, refactored)
 
+  // Insert a new function.
   const args: Ast.Ast[] = info.extracted.inputs.map((arg) => Ast.Ident.new(edit, arg))
-  const collapsedFunction = Ast.Function.new(edit, collapsedName, args, collapsed, true)
+  const collapsedFunction = Ast.Function.fromExprs(edit, collapsedName, args, collapsed, true)
   topLevel.insert(edit, posToInsert, collapsedFunction)
-  return edit
+  return { refactoredNodeId, collapsedNodeIds, outputNodeId }
 }
 
 /** Prepare a method call expression for collapsed method. */
@@ -188,13 +208,14 @@ function collapsedCallAst(
   info: CollapsedInfo,
   collapsedName: string,
   edit: Ast.MutableModule,
-): Ast.Ast {
+): { ast: Ast.Ast; nodeId: ExprId } {
   const pattern = info.refactored.pattern
   const args = info.refactored.arguments
   const functionName = `${MODULE_NAME}.${collapsedName}`
   const expression = functionName + (args.length > 0 ? ' ' : '') + args.join(' ')
-  const assignment = Ast.Assignment.new(edit, pattern, Ast.parse(expression, edit))
-  return assignment
+  const expressionAst = Ast.parse(expression, edit)
+  const ast = Ast.Assignment.new(edit, pattern, expressionAst)
+  return { ast, nodeId: expressionAst.exprId }
 }
 
 /** Find the position before the current method to insert a collapsed one. */