+++ app/ide-desktop/lib/dashboard/src/pages/dashboard/Dashboard.tsx
@@ -31,7 +31,6 @@ import * as assetQuery from '#/utilities/assetQuery'
 import * as http from '#/utilities/http'
 import * as localStorageModule from '#/utilities/localStorage'
 import * as object from '#/utilities/object'
-import * as projectManager from '#/utilities/projectManager'
 import * as shortcutsModule from '#/utilities/shortcuts'
 
 import TheModal from '#/components/dashboard/TheModal'
@@ -65,7 +64,6 @@ export default function Dashboard(props: DashboardProps) {
     const [initialized, setInitialized] = React.useState(false)
     const [isHelpChatOpen, setIsHelpChatOpen] = React.useState(false)
     const [isHelpChatVisible, setIsHelpChatVisible] = React.useState(false)
-    const [loadingProjectManagerDidFail, setLoadingProjectManagerDidFail] = React.useState(false)
     const [page, setPage] = React.useState(
         () => localStorage.get(localStorageModule.LocalStorageKey.page) ?? pageSwitcher.Page.drive
     )
@@ -92,15 +90,6 @@ export default function Dashboard(props: DashboardProps) {
         () => session.organization?.rootDirectoryId ?? backendModule.DirectoryId(''),
         [session.organization]
     )
-
-    const isListingLocalDirectoryAndWillFail =
-        backend.type === backendModule.BackendType.local && loadingProjectManagerDidFail
-    const isListingRemoteDirectoryAndWillFail =
-        backend.type === backendModule.BackendType.remote &&
-        session.organization?.isEnabled !== true
-    const isListingRemoteDirectoryWhileOffline =
-        session.type === authProvider.UserSessionType.offline &&
-        backend.type === backendModule.BackendType.remote
 
     React.useEffect(() => {
         setInitialized(true)
@@ -289,22 +278,6 @@ export default function Dashboard(props: DashboardProps) {
     }, [isHelpChatOpen])
 
     React.useEffect(() => {
-        const onProjectManagerLoadingFailed = () => {
-            setLoadingProjectManagerDidFail(true)
-        }
-        document.addEventListener(
-            projectManager.ProjectManagerEvents.loadingFailed,
-            onProjectManagerLoadingFailed
-        )
-        return () => {
-            document.removeEventListener(
-                projectManager.ProjectManagerEvents.loadingFailed,
-                onProjectManagerLoadingFailed
-            )
-        }
-    }, [])
-
-    React.useEffect(() => {
         return shortcuts.registerKeyboardHandlers({
             [shortcutsModule.KeyboardAction.closeModal]: () => {
                 unsetModal()
@@ -344,15 +317,17 @@ export default function Dashboard(props: DashboardProps) {
 
     const doCreateProject = React.useCallback(
         (
-            templateId: string | null,
-            onSpinnerStateChange?: (state: spinner.SpinnerState) => void
+            templateId: string | null = null,
+            templateName: string | null = null,
+            onSpinnerStateChange: ((state: spinner.SpinnerState) => void) | null = null
         ) => {
             dispatchAssetListEvent({
                 type: AssetListEventType.newProject,
                 parentKey: rootDirectoryId,
                 parentId: rootDirectoryId,
-                templateId: templateId ?? null,
-                onSpinnerStateChange: onSpinnerStateChange ?? null,
+                templateId: templateId,
+                templateName: templateName,
+                onSpinnerStateChange: onSpinnerStateChange,
             })
         },
         [rootDirectoryId, /* should never change */ dispatchAssetListEvent]
@@ -446,7 +421,7 @@ export default function Dashboard(props: DashboardProps) {
                     />
                     <Home
                         hidden={page !== pageSwitcher.Page.home}
-                        onTemplateClick={doCreateProject}
+                        createProject={doCreateProject}
                     />
                     <Drive
                         supportsLocalBackend={supportsLocalBackend}
@@ -468,10 +443,6 @@ export default function Dashboard(props: DashboardProps) {
                         doCreateProject={doCreateProject}
                         doOpenEditor={openEditor}
                         doCloseEditor={closeEditor}
-                        loadingProjectManagerDidFail={loadingProjectManagerDidFail}
-                        isListingRemoteDirectoryWhileOffline={isListingRemoteDirectoryWhileOffline}
-                        isListingLocalDirectoryAndWillFail={isListingLocalDirectoryAndWillFail}
-                        isListingRemoteDirectoryAndWillFail={isListingRemoteDirectoryAndWillFail}
                     />
                     <Editor
                         hidden={page !== pageSwitcher.Page.editor}