+++ CHANGELOG.md
@@ -599,6 +599,7 @@
 - [Support for loading big Excel files.][8403]
 - [Added new `Filter_Condition`s - `Equal_Ignore_Case`, `Is_Nan`, `Is_Infinite`
   and `Is_Finite`.][8539]
+- [Support `on_problems=Problem_Behavior.Report_Warning` in `Vector.map`.][8595]
 - [Added text_length to Column][8606]
 - [Added none delimiter option for Data.Read][8627]
 - [Added text_left and text_right to Column][8691]
@@ -861,6 +862,7 @@
 [8403]: https://github.com/enso-org/enso/pull/8403
 [8539]: https://github.com/enso-org/enso/pull/8539
 [8564]: https://github.com/enso-org/enso/pull/8564
+[8595]: https://github.com/enso-org/enso/pull/8595
 [8606]: https://github.com/enso-org/enso/pull/8606
 [8627]: https://github.com/enso-org/enso/pull/8627
 [8691]: https://github.com/enso-org/enso/pull/8691
@@ -1018,6 +1020,7 @@
 - [Added opt-in type checks of return type][8502]
 - [Introduce Arrow language][8512]
 - [DataflowError.withoutTrace doesn't store stacktrace][8608]
+- [Derive --in-project from --run source location][8775]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -1171,6 +1174,7 @@
 [8502]: https://github.com/enso-org/enso/pull/8502
 [8512]: https://github.com/enso-org/enso/pull/8512
 [8608]: https://github.com/enso-org/enso/pull/8608
+[8775]: https://github.com/enso-org/enso/pull/8775
 
 # Enso 2.0.0-alpha.18 (2021-10-12)